weird sample pos interpolation bug fixed

+++ include/texture.h
@@ -5,7 +5,7 @@
 
 void create_texture3D(
     i32 width, i32 height, i32 depth,
-    u8 (*function)(float x, float y, float z),
+    float (*function)(float x, float y, float z),
     u32* texture
 );
